fmpz_poly_q scalar
